Make image use cert-manager to obtain secrets
I love [cert-manager](https://github.com/jetstack/cert-manager). They
really are close to having an ideal solution to issuing certs in a
kubernetes cluster. I think they are closer to 1.0 then they are aware.

Cert-manager has a "CA Issuer" which, when provided with a signing
key/cert will use them to create and sign certs for us. While these
certs are not useful for general consumption, it is perfect for
encryption and identification within the cluster. This should use that
facility to encrypt the DB traffic.

Useing a cert to identify the client is out of scope for this commit,
but should be a fairly easy improvement later on.